fully implemented syncing gold:

+++ Assets/scripts/Globals.cs
@@ -3,6 +3,6 @@ using System.Collections;
 
 public class Globals {
 	public static AddState addState { get; set; }
-	public static float gold { get; set; }
-	public static float goldPerSec { get; set; }
+	public static double gold { get; set; }
+	public static double goldPerSec { get; set; }
 }
